Added example of using Shape as a clip, with interactionSource.
